Initial work on allowing different evaluators.
