+++ lib/jquery_query_builder/evaluator.rb
@@ -1,30 +1,7 @@
-require 'active_support/core_ext/hash/indifferent_access'
-require 'active_support/core_ext/object/blank'
-require 'active_support/core_ext/string/inflections'
-require 'jquery_query_builder/operator'
-require 'jquery_query_builder/rule_group'
-require 'jquery_query_builder/rule'
-require 'json'
+require 'jquery_query_builder/evaluators/boolean'
 
 module JqueryQueryBuilder
-  class Evaluator
-    attr_accessor :parsed_rule_set
-    def initialize(rule_set)
-      if rule_set.is_a? String
-        #assuming the json was passed in
-        self.parsed_rule_set = JSON.parse(rule_set)
-      else
-        self.parsed_rule_set = rule_set
-      end
-    end
-
-    def get_matching_objects(objects)
-      objects.select{|o| object_matches_rules?(o)}
-    end
-
-    def object_matches_rules?(object)
-      RuleGroup.new(parsed_rule_set).evaluate(object)
-    end
+  class Evaluator < Evaluators::Boolean
   end
 end
