+++ lib/jquery_query_builder/evaluators/boolean/operators/begins_with.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class BeginsWith
+          def evaluate(left, right)
+            left.start_with?(right)
+          end
+        end
+      end
+    end
+  end
+end
