+++ lib/jquery_query_builder/evaluators/boolean/operators/between.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class Between
+          def evaluate(input, bounds)
+            input > bounds[0] && input < bounds[1]
+          end
+        end
+      end
+    end
+  end
+end
