+++ lib/jquery_query_builder/evaluators/boolean/operators/contains.rb
@@ -1,0 +1,14 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class Contains
+          def evaluate(left, right)
+            return false if left.nil?
+            left.include?(right)
+          end
+        end
+      end
+    end
+  end
+end
