+++ lib/jquery_query_builder/evaluators/boolean/operators/ends_with.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class EndsWith
+          def evaluate(left, right)
+            left.end_with?(right)
+          end
+        end
+      end
+    end
+  end
+end
