+++ lib/jquery_query_builder/evaluators/boolean/operators/greater.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class Greater
+          def evaluate(left, right)
+            left > right
+          end
+        end
+      end
+    end
+  end
+end
