+++ lib/jquery_query_builder/evaluators/boolean/operators/greater_or_equal.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class GreaterOrEqual
+          def evaluate(left, right)
+            left >= right
+          end
+        end
+      end
+    end
+  end
+end
