+++ lib/jquery_query_builder/evaluators/boolean/operators/is_empty.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class IsEmpty
+          def evaluate(input, value)
+            input.blank?
+          end
+        end
+      end
+    end
+  end
+end
