+++ lib/jquery_query_builder/evaluators/boolean/operators/is_not_empty.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class IsNotEmpty
+          def evaluate(input, value)
+            input.present?
+          end
+        end
+      end
+    end
+  end
+end
