+++ lib/jquery_query_builder/evaluators/boolean/operators/is_not_null.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class IsNotNull
+          def evaluate(input, value)
+            !input.nil?
+          end
+        end
+      end
+    end
+  end
+end
