+++ lib/jquery_query_builder/evaluators/boolean/operators/less.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class Less
+          def evaluate(left, right)
+            left < right
+          end
+        end
+      end
+    end
+  end
+end
