+++ lib/jquery_query_builder/evaluators/boolean/operators/not_contains.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class NotContains
+          def evaluate(left, right)
+            !left.include?(right)
+          end
+        end
+      end
+    end
+  end
+end
