+++ lib/jquery_query_builder/evaluators/boolean/operators/not_in.rb
@@ -1,0 +1,13 @@
+module JqueryQueryBuilder
+  module Evaluators
+    module Boolean
+      module Operators
+        class NotIn
+          def evaluate(left, right)
+            !right.include?(left)
+          end
+        end
+      end
+    end
+  end
+end
