+++ lib/jquery_query_builder/evaluators/sql.rb
@@ -1,0 +1,7 @@
+module JqueryQueryBuilder
+  module Evaluators
+    class Sql < JqueryQueryBuilder::Evaluator
+      
+    end
+  end
+end
