--- lib/jquery_query_builder/operators/between.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class Between
-      def evaluate(input, bounds)
-        input > bounds[0] && input < bounds[1]
-      end
-    end
-  end
-end
