--- lib/jquery_query_builder/operators/contains.rb
@@ -1,10 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class Contains
-      def evaluate(left, right)
-        return false if left.nil?
-        left.include?(right)
-      end
-    end
-  end
-end
