--- lib/jquery_query_builder/operators/ends_with.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class EndsWith
-      def evaluate(left, right)
-        left.end_with?(right)
-      end
-    end
-  end
-end
