--- lib/jquery_query_builder/operators/equal.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class Equal
-      def evaluate(left, right)
-        left == right
-      end
-    end
-  end
-end
