--- lib/jquery_query_builder/operators/greater.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class Greater
-      def evaluate(left, right)
-        left > right
-      end
-    end
-  end
-end
