--- lib/jquery_query_builder/operators/greater_or_equal.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class GreaterOrEqual
-      def evaluate(left, right)
-        left >= right
-      end
-    end
-  end
-end
