--- lib/jquery_query_builder/operators/is_empty.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class IsEmpty
-      def evaluate(input, value)
-        input.blank?
-      end
-    end
-  end
-end
