--- lib/jquery_query_builder/operators/is_not_empty.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class IsNotEmpty
-      def evaluate(input, value)
-        input.present?
-      end
-    end
-  end
-end
