--- lib/jquery_query_builder/operators/is_not_null.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class IsNotNull
-      def evaluate(input, value)
-        !input.nil?
-      end
-    end
-  end
-end
