--- lib/jquery_query_builder/operators/is_null.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class IsNull
-      def evaluate(input, value)
-        input.nil?
-      end
-    end
-  end
-end
