--- lib/jquery_query_builder/operators/less.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class Less
-      def evaluate(left, right)
-        left < right
-      end
-    end
-  end
-end
