--- lib/jquery_query_builder/operators/less_or_equal.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class LessOrEqual
-      def evaluate(left, right)
-        left <= right
-      end
-    end
-  end
-end
