--- lib/jquery_query_builder/operators/not_begins_with.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class NotBeginsWith
-      def evaluate(left, right)
-        !left.start_with?(right)
-      end
-    end
-  end
-end
