--- lib/jquery_query_builder/operators/not_between.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class NotBetween
-      def evaluate(input, bounds)
-        input <= bounds[0] || input >= bounds[1]
-      end
-    end
-  end
-end
