--- lib/jquery_query_builder/operators/not_contains.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class NotContains
-      def evaluate(left, right)
-        !left.include?(right)
-      end
-    end
-  end
-end
