--- lib/jquery_query_builder/operators/not_equal.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class NotEqual
-      def evaluate(left, right)
-        left != right
-      end
-    end
-  end
-end
