--- lib/jquery_query_builder/operators/not_in.rb
@@ -1,9 +1,0 @@
-module JqueryQueryBuilder
-  module Operators
-    class NotIn
-      def evaluate(left, right)
-        !right.include?(left)
-      end
-    end
-  end
-end
